Test the threading model of the AllJoyn framework.
1. Spin more than one thread to call the AllJoyn method at the same time. Note that
although the concurrency of AboutService is set to the default value 4, the AllJoyn
method call is serialized at the BusObject of AboutService.

2. If we increase the number of threads which call the AllJoyn method to 1000, we start
observing many failures for the IntrospectRemoteObject call of the ProxyBusObject with
the error code ER_BUS_REPLY_IS_ERROR_MESSAGE. In this case, increasing the concurrency
of AboutService may reduce the occurrences of the failures but it can't eliminate the
failures.