Update the project settings to specify the build and link dependencies.
Previously we used the hard-coded relative paths and library names to specify
the build and link dependencies of the sample console projects (e.g., control
panel controllee and controller services) on the SampleCommon project and the
Generated&Provided project. With this change, we are using the Visual Studio
project settings to specify those build and link dependencies.
(1) Go to the "PROJECT" menu and open "Project Dependencies". Then specify the
build dependencies.
(2) Right click on each sample console project, select "Properties->Common Properties
->Framework and References->Add New Reference" and add the static library project;
also check that "Link Library Dependencies" is True on the right hand side.
For details, please refer to
http://stackoverflow.com/questions/3795567/visual-studio-2010-not-autolinking-static-libraries-from-projects-that-are-depend.